requirements.txt
Package                      Version
---------------------------- ---------
annotated-types              0.7.0
cachetools                   5.5.0
certifi                      2024.8.30
charset-normalizer           3.3.2
google-ai-generativelanguage 0.6.6
google-api-core              2.19.2
google-api-python-client     2.143.0
google-auth                  2.34.0
google-auth-httplib2         0.2.0
google-generativeai          0.7.2
googleapis-common-protos     1.65.0
grpcio                       1.66.1
grpcio-status                1.62.3
httplib2                     0.22.0
idna                         3.8
pip                          24.2
proto-plus                   1.24.0
protobuf                     4.25.4
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pydantic                     2.8.2
pydantic_core                2.20.1
pyparsing                    3.1.4
requests                     2.32.3
rsa                          4.9
tqdm                         4.66.5
typing_extensions            4.12.2
uritemplate                  4.1.1
urllib3                      2.2.2
